minor refactoring in products
