+++ .vscode/settings.json
@@ -7,5 +7,8 @@
   },
   "files.associations": {
     "*.css": "tailwindcss"
+  },
+  "[typescript]": {
+    "editor.defaultFormatter": "vscode.typescript-language-features"
   }
 }
