+++ packages/backend/src/repositories/products/products.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
-import { ProductsService } from './products.service';
+import { ProductsService } from './products.repository';
 
 @Controller('products')
 export class ProductsController {
