+++ packages/backend/src/repositories/products/products.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ProductsService } from './products.service';
 import { ProductsController } from './products.controller';
+import { ProductsService } from './products.repository';
 
 @Module({
   controllers: [ProductsController],
